+++ npm/designsystem/src/components/Input/Input.tsx
@@ -121,7 +121,7 @@ const Input = React.forwardRef((props: InputProps, ref: React.Ref<HTMLInputEleme
     testId,
     disabled,
     readOnly,
-    autoComplete,
+    autoComplete = 'off',
     afterInputChildren,
     rightOfInput,
     width,